bdgbroadcall

Call broad peaks from bedGraph output. Note: All regions on the same chromosome in the bedGraph file should be continuous so only bedGraph files from MACS3 are accpetable.

bdgbroadcall(
  ifile,
  cutoffpeak = 2,
  cutofflink = 1,
  minlen = 200L,
  lvl1maxgap = 30L,
  lvl2maxgap = 800L,
  trackline = TRUE,
  outdir = ".",
  outputfile = character(),
  log = TRUE,
  verbose = 2L
)

Arguments

ifile

MACS score in bedGraph. REQUIRED.

cutoffpeak

Cutoff for peaks depending on which method you used for score track. If the file contains qvalue scores from MACS3, score 2 means qvalue 0.01. DEFAULT: 2

cutofflink

Cutoff for linking regions/low abundance regions depending on which method you used for score track. If the file contains qvalue scores from MACS3, score 1 means qvalue 0.1, and score 0.3 means qvalue 0.5. DEFAULT: 1", default = 1

minlen

minimum length of peak, better to set it as d value. DEFAULT: 200", default = 200

lvl1maxgap

maximum gap between significant peaks, better to set it as tag size. DEFAULT: 30

lvl2maxgap

maximum linking between significant peaks, better to set it as 4 times of d value. DEFAULT: 800

trackline

Tells MACS not to include trackline with bedGraph files. The trackline is required by UCSC.

outdir

The output directory.

outputfile

The output file.

log

Whether to capture logs.

verbose

Set verbose level of runtime message. 0: only show critical message, 1: show additional warning message, 2: show process information, 3: show debug messages. DEFAULT:2

Value

macsList object.

Examples

eh <- ExperimentHub::ExperimentHub()
CHIP <- eh[["EH4558"]]
#> see ?MACSdata and browseVignettes('MACSdata') for documentation
#> loading from cache
CTRL <- eh[["EH4563"]]
#> see ?MACSdata and browseVignettes('MACSdata') for documentation
#> loading from cache
p1 <- pileup(CHIP, outdir = tempdir(),
             outputfile = "pileup_ChIP_bed.bdg", format = "BED")
#> INFO  @ 17 Nov 2023 12:40:58: [691 MB] # read alignment files... 
#> INFO  @ 17 Nov 2023 12:40:58: [691 MB] # read treatment tags... 
#> INFO  @ 17 Nov 2023 12:40:58: [698 MB] tag size is determined as 101 bps 
#> INFO  @ 17 Nov 2023 12:40:58: [698 MB] # tag size = 101 
#> INFO  @ 17 Nov 2023 12:40:58: [698 MB] # total tags in alignment file: 49622 
#> INFO  @ 17 Nov 2023 12:40:58: [698 MB] # Pileup alignment file, extend each read towards downstream direction with 200 bps 
#> INFO  @ 17 Nov 2023 12:40:58: [698 MB] # Done! Check pileup_ChIP_bed.bdg 
#> 
p2 <- pileup(CTRL, outdir = tempdir(),
             outputfile = "pileup_CTRL_bed.bdg", format = "BED")
#> INFO  @ 17 Nov 2023 12:40:58: [698 MB] # read alignment files... 
#> INFO  @ 17 Nov 2023 12:40:58: [698 MB] # read treatment tags... 
#> INFO  @ 17 Nov 2023 12:40:58: [698 MB] tag size is determined as 101 bps 
#> INFO  @ 17 Nov 2023 12:40:58: [698 MB] # tag size = 101 
#> INFO  @ 17 Nov 2023 12:40:58: [698 MB] # total tags in alignment file: 50837 
#> INFO  @ 17 Nov 2023 12:40:58: [698 MB] # Pileup alignment file, extend each read towards downstream direction with 200 bps 
#> INFO  @ 17 Nov 2023 12:40:58: [698 MB] # Done! Check pileup_CTRL_bed.bdg 
#> 
c1 <- bdgcmp(p1$outputs, p2$outputs, outdir = tempdir(),
             oprefix = "bdgcmp", pseudocount = 1, method = "FE")
#> INFO  @ 17 Nov 2023 12:40:58: [698 MB] Read and build treatment bedGraph... 
#> INFO  @ 17 Nov 2023 12:40:58: [698 MB] Read and build control bedGraph... 
#> INFO  @ 17 Nov 2023 12:40:59: [698 MB] Build ScoreTrackII... 
#> INFO  @ 17 Nov 2023 12:40:59: [700 MB] Calculate scores comparing treatment and control by 'FE'... 
#> INFO  @ 17 Nov 2023 12:41:00: [700 MB] Write bedGraph of scores... 
#> INFO  @ 17 Nov 2023 12:41:00: [700 MB] Finished 'FE'! Please check '/tmp/RtmpeMHyqk/bdgcmp_FE.bdg'! 
#> 
bdgbroadcall(c1$outputs, cutoffpeak = 2, cutofflink = 1.5,
             outdir = tempdir(), outputfile = "bdgbroadcall")
#> INFO  @ 17 Nov 2023 12:41:00: [700 MB] Read and build bedGraph... 
#> INFO  @ 17 Nov 2023 12:41:01: [700 MB] Call peaks from bedGraph... 
#> INFO  @ 17 Nov 2023 12:41:01: [702 MB] Write peaks... 
#> INFO  @ 17 Nov 2023 12:41:01: [702 MB] Done 
#> 
#> macsList class
#> $outputs:
#>  /tmp/RtmpeMHyqk/bdgbroadcall 
#> $arguments: ifile, cutoffpeak, cutofflink, outdir, outputfile 
#> $log:
#>  INFO  @ 17 Nov 2023 12:41:00: [700 MB] Read and build bedGraph... 
#>  INFO  @ 17 Nov 2023 12:41:01: [700 MB] Call peaks from bedGraph... 
#>  INFO  @ 17 Nov 2023 12:41:01: [702 MB] Write peaks... 
#>  INFO  @ 17 Nov 2023 12:41:01: [702 MB] Done 
#>