bdgdiff

Differential peak detection based on paired four bedgraph files. Note: All regions on the same chromosome in the bedGraph file should be continuous so only bedGraph files from MACS3 are accpetable.

bdgdiff(
  t1bdg,
  t2bdg,
  c1bdg,
  c2bdg,
  cutoff = 3,
  minlen = 200L,
  maxgap = 100L,
  depth1 = 1,
  depth2 = 1,
  outdir = ".",
  oprefix = character(),
  outputfile = list(),
  log = TRUE,
  verbose = 2L
)

Arguments

t1bdg

MACS pileup bedGraph for condition 1. Incompatible with callpeak --SPMR output. REQUIRED

t2bdg

MACS pileup bedGraph for condition 2. Incompatible with callpeak --SPMR output. REQUIRED

c1bdg

MACS control lambda bedGraph for condition 1. Incompatible with callpeak --SPMR output. REQUIRED

c2bdg

MACS control lambda bedGraph for condition 2. Incompatible with callpeak --SPMR output. REQUIRED

cutoff

logLR cutoff. DEFAULT: 3 (likelihood ratio=1000)", default = 3

minlen

Minimum length of differential region. Try bigger value to remove small regions. DEFAULT: 200", default = 200

maxgap

Maximum gap to merge nearby differential regions. Consider a wider gap for broad marks. Maximum gap should be smaller than minimum length (-g). DEFAULT: 100", default = 100

depth1

Sequencing depth (# of non-redundant reads in million) for condition 1. It will be used together with --d2. See description for --d2 below for how to assign them. Default: 1

depth2

Sequencing depth (# of non-redundant reads in million) for condition 2. It will be used together with --d1. DEPTH1 and DEPTH2 will be used to calculate scaling factor for each sample, to down-scale larger sample to the level of smaller one. For example, while comparing 10 million condition 1 and 20 million condition 2, use --d1 10 --d2 20, then pileup value in bedGraph for condition 2 will be divided by 2. Default: 1

outdir

The output directory.

oprefix

Output file prefix. Actual files will be named as PREFIX_cond1.bed, PREFIX_cond2.bed and PREFIX_common.bed. Mutually exclusive with -o/--ofile.

outputfile

Output filenames. Must give three arguments in order: 1. file for unique regions in condition 1; 2. file for unique regions in condition 2; 3. file for common regions in both conditions. Note: mutually exclusive with --o-prefix.

log

Whether to capture logs.

verbose

Set verbose level of runtime message. 0: only show critical message, 1: show additional warning message, 2: show process information, 3: show debug messages. DEFAULT:2

Value

macsList object.

Examples

eh <- ExperimentHub::ExperimentHub()
CHIP <- eh[["EH4558"]]
#> see ?MACSdata and browseVignettes('MACSdata') for documentation
#> loading from cache
CTRL <- eh[["EH4563"]]
#> see ?MACSdata and browseVignettes('MACSdata') for documentation
#> loading from cache
c1 <- callpeak(CHIP, CTRL, gsize = 5.2e7, cutoff_analysis = TRUE,
               outdir = tempdir(), name = "callpeak_narrow0", store_bdg = TRUE)
#> INFO  @ 17 Nov 2023 13:42:13: [729 MB] 
#> # Command line: 
#> # ARGUMENTS LIST:
#> # name = callpeak_narrow0
#> # format = AUTO
#> # ChIP-seq file = ['/home/qhu/.cache/R/ExperimentHub/1071e4528bfa6_4601']
#> # control file = ['/home/qhu/.cache/R/ExperimentHub/1071e71fe08e5_4606']
#> # effective genome size = 5.20e+07
#> # band width = 300
#> # model fold = [5.0, 50.0]
#> # qvalue cutoff = 5.00e-02
#> # The maximum gap between significant sites is assigned as the read length/tag size.
#> # The minimum length of peaks is assigned as the predicted fragment length "d".
#> # Larger dataset will be scaled towards smaller dataset.
#> # Range for calculating regional lambda is: 1000 bps and 10000 bps
#> # Broad region calling is off
#> # Additional cutoff on fold-enrichment is: 0.10
#> # Paired-End mode is off
#>  
#> INFO  @ 17 Nov 2023 13:42:13: [729 MB] #1 read tag files... 
#> INFO  @ 17 Nov 2023 13:42:13: [729 MB] #1 read treatment tags... 
#> INFO  @ 17 Nov 2023 13:42:13: [732 MB] Detected format is: BED 
#> INFO  @ 17 Nov 2023 13:42:13: [732 MB] * Input file is gzipped. 
#> INFO  @ 17 Nov 2023 13:42:13: [733 MB] #1.2 read input tags... 
#> INFO  @ 17 Nov 2023 13:42:13: [733 MB] Detected format is: BED 
#> INFO  @ 17 Nov 2023 13:42:13: [733 MB] * Input file is gzipped. 
#> INFO  @ 17 Nov 2023 13:42:13: [734 MB] #1 tag size is determined as 101 bps 
#> INFO  @ 17 Nov 2023 13:42:13: [734 MB] #1 tag size = 101.0 
#> INFO  @ 17 Nov 2023 13:42:13: [734 MB] #1  total tags in treatment: 49622 
#> INFO  @ 17 Nov 2023 13:42:13: [734 MB] #1 user defined the maximum tags... 
#> INFO  @ 17 Nov 2023 13:42:13: [734 MB] #1 filter out redundant tags at the same location and the same strand by allowing at most 1 tag(s) 
#> INFO  @ 17 Nov 2023 13:42:13: [734 MB] #1  tags after filtering in treatment: 48047 
#> INFO  @ 17 Nov 2023 13:42:13: [734 MB] #1  Redundant rate of treatment: 0.03 
#> INFO  @ 17 Nov 2023 13:42:13: [734 MB] #1  total tags in control: 50837 
#> INFO  @ 17 Nov 2023 13:42:13: [734 MB] #1 user defined the maximum tags... 
#> INFO  @ 17 Nov 2023 13:42:13: [734 MB] #1 filter out redundant tags at the same location and the same strand by allowing at most 1 tag(s) 
#> INFO  @ 17 Nov 2023 13:42:13: [734 MB] #1  tags after filtering in control: 50783 
#> INFO  @ 17 Nov 2023 13:42:13: [734 MB] #1  Redundant rate of control: 0.00 
#> INFO  @ 17 Nov 2023 13:42:13: [734 MB] #1 finished! 
#> INFO  @ 17 Nov 2023 13:42:13: [734 MB] #2 Build Peak Model... 
#> INFO  @ 17 Nov 2023 13:42:13: [734 MB] #2 looking for paired plus/minus strand peaks... 
#> INFO  @ 17 Nov 2023 13:42:13: [734 MB] #2 Total number of paired peaks: 469 
#> INFO  @ 17 Nov 2023 13:42:13: [734 MB] #2 Model building with cross-correlation: Done 
#> INFO  @ 17 Nov 2023 13:42:13: [734 MB] #2 finished! 
#> INFO  @ 17 Nov 2023 13:42:13: [734 MB] #2 predicted fragment length is 228 bps 
#> INFO  @ 17 Nov 2023 13:42:13: [734 MB] #2 alternative fragment length(s) may be 228 bps 
#> INFO  @ 17 Nov 2023 13:42:13: [734 MB] #2.2 Generate R script for model : /tmp/RtmpEPj3Vo/callpeak_narrow0_model.r 
#> INFO  @ 17 Nov 2023 13:42:13: [734 MB] #3 Call peaks... 
#> INFO  @ 17 Nov 2023 13:42:13: [734 MB] #3 Pre-compute pvalue-qvalue table... 
#> INFO  @ 17 Nov 2023 13:42:13: [734 MB] #3 Cutoff vs peaks called will be analyzed! 
#> INFO  @ 17 Nov 2023 13:42:14: [757 MB] #3 Analysis of cutoff vs num of peaks or total length has been saved in b'/tmp/RtmpEPj3Vo/callpeak_narrow0_cutoff_analysis.txt' 
#> INFO  @ 17 Nov 2023 13:42:14: [757 MB] #3 In the peak calling step, the following will be performed simultaneously: 
#> INFO  @ 17 Nov 2023 13:42:14: [757 MB] #3   Write bedGraph files for treatment pileup (after scaling if necessary)... callpeak_narrow0_treat_pileup.bdg 
#> INFO  @ 17 Nov 2023 13:42:14: [757 MB] #3   Write bedGraph files for control lambda (after scaling if necessary)... callpeak_narrow0_control_lambda.bdg 
#> INFO  @ 17 Nov 2023 13:42:14: [757 MB] #3   Pileup will be based on sequencing depth in treatment. 
#> INFO  @ 17 Nov 2023 13:42:14: [757 MB] #3 Call peaks for each chromosome... 
#> INFO  @ 17 Nov 2023 13:42:14: [757 MB] #4 Write output xls file... /tmp/RtmpEPj3Vo/callpeak_narrow0_peaks.xls 
#> INFO  @ 17 Nov 2023 13:42:14: [757 MB] #4 Write peak in narrowPeak format file... /tmp/RtmpEPj3Vo/callpeak_narrow0_peaks.narrowPeak 
#> INFO  @ 17 Nov 2023 13:42:14: [757 MB] #4 Write summits bed file... /tmp/RtmpEPj3Vo/callpeak_narrow0_summits.bed 
#> INFO  @ 17 Nov 2023 13:42:14: [757 MB] Done! 
#> 
c2 <- callpeak(CHIP, CTRL, gsize = 1e7, nomodel = TRUE, extsize = 250,
               outdir = tempdir(), name = "callpeak_narrow_revert", store_bdg = TRUE)
#> INFO  @ 17 Nov 2023 13:42:14: [757 MB] 
#> # Command line: 
#> # ARGUMENTS LIST:
#> # name = callpeak_narrow_revert
#> # format = AUTO
#> # ChIP-seq file = ['/home/qhu/.cache/R/ExperimentHub/1071e4528bfa6_4601']
#> # control file = ['/home/qhu/.cache/R/ExperimentHub/1071e71fe08e5_4606']
#> # effective genome size = 1.00e+07
#> # band width = 300
#> # model fold = [5.0, 50.0]
#> # qvalue cutoff = 5.00e-02
#> # The maximum gap between significant sites is assigned as the read length/tag size.
#> # The minimum length of peaks is assigned as the predicted fragment length "d".
#> # Larger dataset will be scaled towards smaller dataset.
#> # Range for calculating regional lambda is: 1000 bps and 10000 bps
#> # Broad region calling is off
#> # Additional cutoff on fold-enrichment is: 0.10
#> # Paired-End mode is off
#>  
#> INFO  @ 17 Nov 2023 13:42:14: [757 MB] #1 read tag files... 
#> INFO  @ 17 Nov 2023 13:42:14: [757 MB] #1 read treatment tags... 
#> INFO  @ 17 Nov 2023 13:42:14: [757 MB] Detected format is: BED 
#> INFO  @ 17 Nov 2023 13:42:14: [757 MB] * Input file is gzipped. 
#> INFO  @ 17 Nov 2023 13:42:14: [758 MB] #1.2 read input tags... 
#> INFO  @ 17 Nov 2023 13:42:14: [758 MB] Detected format is: BED 
#> INFO  @ 17 Nov 2023 13:42:14: [758 MB] * Input file is gzipped. 
#> INFO  @ 17 Nov 2023 13:42:14: [758 MB] #1 tag size is determined as 101 bps 
#> INFO  @ 17 Nov 2023 13:42:14: [758 MB] #1 tag size = 101.0 
#> INFO  @ 17 Nov 2023 13:42:14: [758 MB] #1  total tags in treatment: 49622 
#> INFO  @ 17 Nov 2023 13:42:14: [758 MB] #1 user defined the maximum tags... 
#> INFO  @ 17 Nov 2023 13:42:14: [758 MB] #1 filter out redundant tags at the same location and the same strand by allowing at most 1 tag(s) 
#> INFO  @ 17 Nov 2023 13:42:14: [758 MB] #1  tags after filtering in treatment: 48047 
#> INFO  @ 17 Nov 2023 13:42:14: [758 MB] #1  Redundant rate of treatment: 0.03 
#> INFO  @ 17 Nov 2023 13:42:14: [758 MB] #1  total tags in control: 50837 
#> INFO  @ 17 Nov 2023 13:42:14: [758 MB] #1 user defined the maximum tags... 
#> INFO  @ 17 Nov 2023 13:42:14: [758 MB] #1 filter out redundant tags at the same location and the same strand by allowing at most 1 tag(s) 
#> INFO  @ 17 Nov 2023 13:42:14: [758 MB] #1  tags after filtering in control: 50783 
#> INFO  @ 17 Nov 2023 13:42:14: [758 MB] #1  Redundant rate of control: 0.00 
#> INFO  @ 17 Nov 2023 13:42:14: [758 MB] #1 finished! 
#> INFO  @ 17 Nov 2023 13:42:14: [758 MB] #2 Build Peak Model... 
#> INFO  @ 17 Nov 2023 13:42:14: [758 MB] #2 Skipped... 
#> INFO  @ 17 Nov 2023 13:42:14: [758 MB] #2 Use 250 as fragment length 
#> INFO  @ 17 Nov 2023 13:42:14: [758 MB] #3 Call peaks... 
#> INFO  @ 17 Nov 2023 13:42:14: [758 MB] #3 Pre-compute pvalue-qvalue table... 
#> INFO  @ 17 Nov 2023 13:42:14: [758 MB] #3 In the peak calling step, the following will be performed simultaneously: 
#> INFO  @ 17 Nov 2023 13:42:14: [758 MB] #3   Write bedGraph files for treatment pileup (after scaling if necessary)... callpeak_narrow_revert_treat_pileup.bdg 
#> INFO  @ 17 Nov 2023 13:42:14: [758 MB] #3   Write bedGraph files for control lambda (after scaling if necessary)... callpeak_narrow_revert_control_lambda.bdg 
#> INFO  @ 17 Nov 2023 13:42:14: [758 MB] #3   Pileup will be based on sequencing depth in treatment. 
#> INFO  @ 17 Nov 2023 13:42:14: [758 MB] #3 Call peaks for each chromosome... 
#> INFO  @ 17 Nov 2023 13:42:14: [758 MB] #4 Write output xls file... /tmp/RtmpEPj3Vo/callpeak_narrow_revert_peaks.xls 
#> INFO  @ 17 Nov 2023 13:42:14: [758 MB] #4 Write peak in narrowPeak format file... /tmp/RtmpEPj3Vo/callpeak_narrow_revert_peaks.narrowPeak 
#> INFO  @ 17 Nov 2023 13:42:14: [758 MB] #4 Write summits bed file... /tmp/RtmpEPj3Vo/callpeak_narrow_revert_summits.bed 
#> INFO  @ 17 Nov 2023 13:42:14: [758 MB] Done! 
#> 
t1bdg <- grep("treat_pileup", c1$outputs, value = TRUE)
c1bdg <- grep("control_lambda", c1$outputs, value = TRUE)
t2bdg <- grep("treat_pileup", c2$outputs, value = TRUE)
c2bdg <- grep("control_lambda", c2$outputs, value = TRUE)
bdgdiff(t1bdg, t2bdg, c1bdg, c2bdg,
        outdir = tempdir(), oprefix = "bdgdiff")
#> INFO  @ 17 Nov 2023 13:42:14: [758 MB] Read and build treatment 1 bedGraph... 
#> INFO  @ 17 Nov 2023 13:42:14: [758 MB] Read and build control 1 bedGraph... 
#> INFO  @ 17 Nov 2023 13:42:14: [758 MB] Read and build treatment 2 bedGraph... 
#> INFO  @ 17 Nov 2023 13:42:15: [758 MB] Read and build control 2 bedGraph... 
#> INFO  @ 17 Nov 2023 13:42:15: [758 MB] Write peaks... 
#> INFO  @ 17 Nov 2023 13:42:15: [758 MB] Done 
#> 
#> macsList class
#> $outputs:
#>  /tmp/RtmpEPj3Vo/bdgdiff_c3.0_common.bed
#>  /tmp/RtmpEPj3Vo/bdgdiff_c3.0_cond1.bed
#>  /tmp/RtmpEPj3Vo/bdgdiff_c3.0_cond2.bed 
#> $arguments: t1bdg, t2bdg, c1bdg, c2bdg, outdir, oprefix 
#> $log:
#>  INFO  @ 17 Nov 2023 13:42:14: [758 MB] Read and build treatment 1 bedGraph... 
#>  INFO  @ 17 Nov 2023 13:42:14: [758 MB] Read and build control 1 bedGraph... 
#>  INFO  @ 17 Nov 2023 13:42:14: [758 MB] Read and build treatment 2 bedGraph... 
#>  INFO  @ 17 Nov 2023 13:42:15: [758 MB] Read and build control 2 bedGraph... 
#>  INFO  @ 17 Nov 2023 13:42:15: [758 MB] Write peaks... 
#> ...