bdgopt

Operations on score column of bedGraph file. Note: All regions on the same chromosome in the bedGraph file should be continuous so only bedGraph files from MACS3 are accpetable.

bdgopt(
  ifile,
  method = c("multiply", "add", "p2q", "max", "min"),
  extraparam = numeric(),
  outputfile = character(),
  outdir = ".",
  log = TRUE
)

Arguments

ifile
MACS score in bedGraph. Note: this must be a bedGraph file covering the ENTIRE genome. REQUIRED

method
Method to modify the score column of bedGraph file. Available choices are: multiply, add, max, min, or p2q. 1) multiply, the EXTRAPARAM is required and will be multiplied to the score column. If you intend to divide the score column by X, use value of 1/X as EXTRAPARAM. 2) add, the EXTRAPARAM is required and will be added to the score column. If you intend to subtract the score column by X, use value of -X as EXTRAPARAM. 3) max, the EXTRAPARAM is required and will take the maximum value between score and the EXTRAPARAM. 4) min, the EXTRAPARAM is required and will take the minimum value between score and the EXTRAPARAM. 5) p2q, this will convert p-value scores to q-value scores using Benjamini-Hochberg process. The EXTRAPARAM is not required. This method assumes the scores are -log10 p-value from MACS3. Any other types of score will cause unexpected errors.", default="p2q"

extraparam
The extra parameter for METHOD. Check the detail of -m option.

outputfile
Output filename. Mutually exclusive with --o-prefix. The number and the order of arguments for --ofile must be the same as for -m.

outdir
The output directory.

log
Whether to capture logs.

Value

macsList object.

Examples

eh <- ExperimentHub::ExperimentHub()

#> snapshotDate(): 2020-12-09

CHIP <- eh[["EH4558"]]

#> see ?MACSdata and browseVignettes('MACSdata') for documentation

#> loading from cache

CTRL <- eh[["EH4563"]]

#> see ?MACSdata and browseVignettes('MACSdata') for documentation

#> loading from cache

c1 <- callpeak(CHIP, CTRL, gsize = 5.2e7, cutoff_analysis = TRUE,
outdir = tempdir(), name = "callpeak_narrow0", store_bdg = TRUE)

#> INFO  @ Sun, 13 Dec 2020 14:23:09: 
#> # Command line: 
#> # ARGUMENTS LIST:
#> # name = callpeak_narrow0
#> # format = AUTO
#> # ChIP-seq file = ['/Users/qi28068/Library/Caches/ExperimentHub/3f8c68423d6b_4601']
#> # control file = ['/Users/qi28068/Library/Caches/ExperimentHub/3f8c54ce7146_4606']
#> # effective genome size = 5.20e+07
#> # band width = 300
#> # model fold = [5.0, 50.0]
#> # qvalue cutoff = 5.00e-02
#> # The maximum gap between significant sites is assigned as the read length/tag size.
#> # The minimum length of peaks is assigned as the predicted fragment length "d".
#> # Larger dataset will be scaled towards smaller dataset.
#> # Range for calculating regional lambda is: 1000 bps and 10000 bps
#> # Broad region calling is off
#> # Additional cutoff on fold-enrichment is: 0.10
#> # Paired-End mode is off
#>  
#> INFO  @ Sun, 13 Dec 2020 14:23:09: #1 read tag files... 
#> INFO  @ Sun, 13 Dec 2020 14:23:09: #1 read treatment tags... 
#> INFO  @ Sun, 13 Dec 2020 14:23:09: Detected format is: BED 
#> INFO  @ Sun, 13 Dec 2020 14:23:09: * Input file is gzipped. 
#> INFO  @ Sun, 13 Dec 2020 14:23:09: #1.2 read input tags... 
#> INFO  @ Sun, 13 Dec 2020 14:23:09: Detected format is: BED 
#> INFO  @ Sun, 13 Dec 2020 14:23:09: * Input file is gzipped. 
#> INFO  @ Sun, 13 Dec 2020 14:23:09: #1 tag size is determined as 101 bps 
#> INFO  @ Sun, 13 Dec 2020 14:23:09: #1 tag size = 101.0 
#> INFO  @ Sun, 13 Dec 2020 14:23:09: #1  total tags in treatment: 49622 
#> INFO  @ Sun, 13 Dec 2020 14:23:09: #1 user defined the maximum tags... 
#> INFO  @ Sun, 13 Dec 2020 14:23:09: #1 filter out redundant tags at the same location and the same strand by allowing at most 1 tag(s) 
#> INFO  @ Sun, 13 Dec 2020 14:23:09: #1  tags after filtering in treatment: 48047 
#> INFO  @ Sun, 13 Dec 2020 14:23:09: #1  Redundant rate of treatment: 0.03 
#> INFO  @ Sun, 13 Dec 2020 14:23:09: #1  total tags in control: 50837 
#> INFO  @ Sun, 13 Dec 2020 14:23:09: #1 user defined the maximum tags... 
#> INFO  @ Sun, 13 Dec 2020 14:23:09: #1 filter out redundant tags at the same location and the same strand by allowing at most 1 tag(s) 
#> INFO  @ Sun, 13 Dec 2020 14:23:09: #1  tags after filtering in control: 50783 
#> INFO  @ Sun, 13 Dec 2020 14:23:09: #1  Redundant rate of control: 0.00 
#> INFO  @ Sun, 13 Dec 2020 14:23:09: #1 finished! 
#> INFO  @ Sun, 13 Dec 2020 14:23:09: #2 Build Peak Model... 
#> INFO  @ Sun, 13 Dec 2020 14:23:09: #2 looking for paired plus/minus strand peaks... 
#> INFO  @ Sun, 13 Dec 2020 14:23:09: #2 Total number of paired peaks: 1036 
#> INFO  @ Sun, 13 Dec 2020 14:23:09: #2 Model building with cross-correlation: Done 
#> INFO  @ Sun, 13 Dec 2020 14:23:09: #2 finished! 
#> INFO  @ Sun, 13 Dec 2020 14:23:09: #2 predicted fragment length is 230 bps 
#> INFO  @ Sun, 13 Dec 2020 14:23:09: #2 alternative fragment length(s) may be 230,238 bps 
#> INFO  @ Sun, 13 Dec 2020 14:23:09: #2.2 Generate R script for model : /var/folders/pv/qm6yts0j7hdgggvq53hrzfr0vfyhvv/T//RtmpsaYg38/callpeak_narrow0_model.r 
#> INFO  @ Sun, 13 Dec 2020 14:23:09: #3 Call peaks... 
#> INFO  @ Sun, 13 Dec 2020 14:23:09: #3 Pre-compute pvalue-qvalue table... 
#> INFO  @ Sun, 13 Dec 2020 14:23:09: #3 Cutoff vs peaks called will be analyzed! 
#> INFO  @ Sun, 13 Dec 2020 14:23:10: #3 Analysis of cutoff vs num of peaks or total length has been saved in b'/var/folders/pv/qm6yts0j7hdgggvq53hrzfr0vfyhvv/T//RtmpsaYg38/callpeak_narrow0_cutoff_analysis.txt' 
#> INFO  @ Sun, 13 Dec 2020 14:23:10: #3 In the peak calling step, the following will be performed simultaneously: 
#> INFO  @ Sun, 13 Dec 2020 14:23:10: #3   Write bedGraph files for treatment pileup (after scaling if necessary)... callpeak_narrow0_treat_pileup.bdg 
#> INFO  @ Sun, 13 Dec 2020 14:23:10: #3   Write bedGraph files for control lambda (after scaling if necessary)... callpeak_narrow0_control_lambda.bdg 
#> INFO  @ Sun, 13 Dec 2020 14:23:10: #3   Pileup will be based on sequencing depth in treatment. 
#> INFO  @ Sun, 13 Dec 2020 14:23:10: #3 Call peaks for each chromosome... 
#> INFO  @ Sun, 13 Dec 2020 14:23:10: #4 Write output xls file... /var/folders/pv/qm6yts0j7hdgggvq53hrzfr0vfyhvv/T//RtmpsaYg38/callpeak_narrow0_peaks.xls 
#> INFO  @ Sun, 13 Dec 2020 14:23:10: #4 Write peak in narrowPeak format file... /var/folders/pv/qm6yts0j7hdgggvq53hrzfr0vfyhvv/T//RtmpsaYg38/callpeak_narrow0_peaks.narrowPeak 
#> INFO  @ Sun, 13 Dec 2020 14:23:10: #4 Write summits bed file... /var/folders/pv/qm6yts0j7hdgggvq53hrzfr0vfyhvv/T//RtmpsaYg38/callpeak_narrow0_summits.bed 
#> INFO  @ Sun, 13 Dec 2020 14:23:10: Done! 
#> 

#> Error in .setupMethodsTables(fdef, initialize = TRUE): trying to get slot "group" from an object of a basic class ("NULL") with no slots

cfile <- grep("treat_pileup.bdg", c1$outputs, value = TRUE)

#> Error in grep("treat_pileup.bdg", c1$outputs, value = TRUE): object 'c1' not found

bdgopt(cfile, method = "min", extraparam = 10,
outdir = tempdir(), outputfile = "bdgopt_min.bdg")

#> Error in file.path(ifile): object 'cfile' not found