bdgpeakcall

Call peaks from bedGraph output. Note: All regions on the same chromosome in the bedGraph file should be continuous so only bedGraph files from MACS3 are accpetable.

bdgpeakcall(
  ifile,
  cutoff = 5,
  minlen = 200L,
  maxgap = 30L,
  call_summits = FALSE,
  cutoff_analysis = FALSE,
  trackline = TRUE,
  outdir = ".",
  outputfile = character(),
  log = TRUE
)

Arguments

ifile
MACS score in bedGraph. REQUIRED.

cutoff
Cutoff depends on which method you used for score track. If the file contains pvalue scores from MACS3, score 5 means pvalue 1e-5. DEFAULT: 5", default = 5.

minlen
minimum length of peak, better to set it as d value. DEFAULT: 200", default = 200.

maxgap
maximum gap between significant points in a peak, better to set it as tag size. DEFAULT: 30", default = 30.

call_summits
If set, MACS will use a more sophisticated approach to find all summits in each enriched peak region DEFAULT: False",default=False.

cutoff_analysis
While set, bdgpeakcall will analyze number or total length of peaks that can be called by different cutoff then output a summary table to help user decide a better cutoff. Note, minlen and maxgap may affect the results. DEFAULT: False", default = False.

trackline
Tells MACS not to include trackline with bedGraph files. The trackline is required by UCSC.

outdir
The output directory.

outputfile
The output file.

log
Whether to capture logs.

Value

macsList object.

Examples

# \donttest{
eh <- ExperimentHub::ExperimentHub()

#> snapshotDate(): 2020-12-09

CHIP <- eh[["EH4558"]]

#> see ?MACSdata and browseVignettes('MACSdata') for documentation

#> loading from cache

CTRL <- eh[["EH4563"]]

#> see ?MACSdata and browseVignettes('MACSdata') for documentation

#> loading from cache

p1 <- pileup(CHIP, outdir = tempdir(),
outputfile = "pileup_ChIP_bed.bdg", format = "BED")

#> INFO  @ Sun, 13 Dec 2020 14:23:14: # Existing file b'/var/folders/pv/qm6yts0j7hdgggvq53hrzfr0vfyhvv/T//RtmpsaYg38/pileup_ChIP_bed.bdg' will be replaced! 
#> INFO  @ Sun, 13 Dec 2020 14:23:14: # read alignment files... 
#> INFO  @ Sun, 13 Dec 2020 14:23:14: # read treatment tags... 
#> INFO  @ Sun, 13 Dec 2020 14:23:14: tag size is determined as 101 bps 
#> INFO  @ Sun, 13 Dec 2020 14:23:14: # tag size = 101 
#> INFO  @ Sun, 13 Dec 2020 14:23:14: # total tags in alignment file: 49622 
#> INFO  @ Sun, 13 Dec 2020 14:23:14: # Pileup alignment file, extend each read towards downstream direction with 200 bps 
#> INFO  @ Sun, 13 Dec 2020 14:23:14: # Done! Check pileup_ChIP_bed.bdg 
#> 

#> Error in .setupMethodsTables(fdef, initialize = TRUE): trying to get slot "group" from an object of a basic class ("NULL") with no slots

p2 <- pileup(CTRL, outdir = tempdir(),
outputfile = "pileup_CTRL_bed.bdg", format = "BED")

#> INFO  @ Sun, 13 Dec 2020 14:23:14: # Existing file b'/var/folders/pv/qm6yts0j7hdgggvq53hrzfr0vfyhvv/T//RtmpsaYg38/pileup_CTRL_bed.bdg' will be replaced! 
#> INFO  @ Sun, 13 Dec 2020 14:23:14: # read alignment files... 
#> INFO  @ Sun, 13 Dec 2020 14:23:14: # read treatment tags... 
#> INFO  @ Sun, 13 Dec 2020 14:23:14: tag size is determined as 101 bps 
#> INFO  @ Sun, 13 Dec 2020 14:23:14: # tag size = 101 
#> INFO  @ Sun, 13 Dec 2020 14:23:14: # total tags in alignment file: 50837 
#> INFO  @ Sun, 13 Dec 2020 14:23:14: # Pileup alignment file, extend each read towards downstream direction with 200 bps 
#> INFO  @ Sun, 13 Dec 2020 14:23:14: # Done! Check pileup_CTRL_bed.bdg 
#> 

#> Error in .setupMethodsTables(fdef, initialize = TRUE): trying to get slot "group" from an object of a basic class ("NULL") with no slots

c1 <- bdgcmp(p1$outputs, p2$outputs, outdir = tempdir(),
oprefix = "bdgcmp", pseudocount = 1, method = "FE")

#> Error in file.path(tfile): object 'p1' not found

bdgpeakcall(c1$outputs, cutoff = 2,
outdir = tempdir(), outputfile = "bdgpeakcall")

#> Error in file.path(ifile): object 'c1' not found

# }