filterdup

filterdup

filterdup(
  ifile,
  gsize = "hs",
  format = "AUTO",
  tsize = NULL,
  pvalue = 1e-05,
  keepduplicates = "auto",
  outputfile = character(),
  outdir = ".",
  verbose = 2L,
  buffer_size = 10000,
  dryrun = FALSE,
  log = TRUE
)

Arguments

ifile

Input file(s).

gsize

Effective genome size. It can be 1.0e+9 or 1000000000, or shortcuts:'hs' for human (2.7e9), 'mm' for mouse (1.87e9), 'ce' for C. elegans (9e7) and 'dm' for fruitfly (1.2e8), Default:hs.

format

Input file format.

tsize

Tag size. This will override the auto detected tag size.

pvalue

Pvalue cutoff for binomial distribution test. DEFAULT:1e-5.

keepduplicates

It controls the behavior towards duplicate tags at the exact same location -- the same coordination and the same strand. The 'auto' option makes MACS calculate the maximum tags at the exact same location based on binomal distribution using 1e-5 as pvalue cutoff; and the 'all' option keeps every tags. If an integer is given, at most this number of tags will be kept at the same location. Note, if you've used samtools or picard to flag reads as 'PCR/Optical duplicate' in bit 1024, MACS2 will still read them although the reads may be decided by MACS2 as duplicate later. If you plan to rely on samtools/picard/any other tool to filter duplicates, please remove those duplicate reads and save a new alignment file then ask MACS2 to keep all by '--keep-dup all'. The default is to keep one tag at the same location. Default: 1".

outputfile

The output file.

outdir

The output directory.

verbose

Set verbose level of runtime message. 0: only show critical message, 1: show additional warning message, 2: show process information, 3: show debug messages. DEFAULT: 2.

buffer_size

Buffer size for incrementally increasing internal array size to store reads alignment information. In most cases, you don't have to change this parameter. However, if there are large number of chromosomes/contigs/scaffolds in your alignment, it's recommended to specify a smaller buffer size in order to decrease memory usage (but it will take longer time to read alignment files). Minimum memory requested for reading an alignment file is about # of CHROMOSOME * BUFFER_SIZE * 8 Bytes. DEFAULT: 100000.

dryrun

When set, filterdup will only output numbers instead of writing output files, including maximum allowable duplicates, total number of reads before filtering, total number of reads after filtering, and redundant rate. Default: not set.

log

Whether to capture logs.

Value

macsList object.

Examples

eh <- ExperimentHub::ExperimentHub()
CHIP <- eh[["EH4558"]]
#> see ?MACSdata and browseVignettes('MACSdata') for documentation
#> loading from cache
res <- filterdup(ifile = CHIP, outputfile = "test.bed", outdir = tempdir())
#> INFO  @ 17 Nov 2023 13:42:36: [761 MB] # read tag files... 
#> INFO  @ 17 Nov 2023 13:42:36: [761 MB] # read treatment tags... 
#> INFO  @ 17 Nov 2023 13:42:36: [761 MB] Detected format is: BED 
#> INFO  @ 17 Nov 2023 13:42:36: [761 MB] * Input file is gzipped. 
#> INFO  @ 17 Nov 2023 13:42:36: [761 MB] # tag size = 101 
#> INFO  @ 17 Nov 2023 13:42:36: [761 MB] # total tags in alignment file: 49622 
#> INFO  @ 17 Nov 2023 13:42:36: [761 MB] calculate max duplicate tags in single position based on binomal distribution... 
#> INFO  @ 17 Nov 2023 13:42:36: [761 MB]  max_dup_tags based on binomal = 1 
#> INFO  @ 17 Nov 2023 13:42:36: [761 MB] filter out redundant tags at the same location and the same strand by allowing at most 1 tag(s) 
#> INFO  @ 17 Nov 2023 13:42:36: [761 MB]  tags after filtering in alignment file: 48047 
#> INFO  @ 17 Nov 2023 13:42:36: [761 MB]  Redundant rate of alignment file: 0.03 
#> INFO  @ 17 Nov 2023 13:42:36: [761 MB] Write to BED file 
#> INFO  @ 17 Nov 2023 13:42:36: [761 MB] finished! Check test.bed. 
#>