predictd

Predict d or fragment size from alignment results. In case of PE data, report the average insertion/fragment size from all pairs. Will NOT filter duplicates

predictd(
  ifile,
  gsize = "hs",
  format = "AUTO",
  plot = file.path(tempdir(), "predictd_mode.pdf"),
  tsize = NULL,
  bw = 300,
  d_min = 20,
  mfold = c(5, 50),
  buffer_size = 1e+05,
  verbose = 2L,
  log = TRUE
)

Arguments

ifile
Input file(s).

gsize
Effective genome size. It can be 1.0e+9 or 1000000000, or shortcuts:'hs' for human (2.7e9), 'mm' for mouse (1.87e9), 'ce' for C. elegans (9e7) and 'dm' for fruitfly (1.2e8), Default:hs.

format
Input file format.

plot
PDF path of peak model and correlation plots.

tsize
Tag size. This will override the auto detected tag size.

bw
Band width for picking regions to compute fragment size. This value is only used while building the shifting model. DEFAULT: 300

d_min
Minimum fragment size in basepair. Any predicted fragment size less than this will be excluded. DEFAULT: 20

mfold
Select the regions within MFOLD range of high-confidence enrichment ratio against background to build model. Fold-enrichment in regions must be lower than upper limit, and higher than the lower limit. Use as "-m 10 30". DEFAULT:5 50

buffer_size
Buffer size for incrementally increasing internal array size to store reads alignment information. DEFAULT: 100000.

verbose
Set verbose level of runtime message. 0: only show critical message, 1: show additional warning message, 2: show process information, 3: show debug messages. DEFAULT:2

log
Whether to capture log.

Value

predicted fragment sizes.

Examples

eh <- ExperimentHub::ExperimentHub()

#> snapshotDate(): 2020-12-09

CHIP <- eh[["EH4558"]]

#> see ?MACSdata and browseVignettes('MACSdata') for documentation

#> loading from cache

predictd(CHIP, d_min = 10, gsize=5.2e+7, plot = NULL)

#> INFO  @ Sun, 13 Dec 2020 14:23:31: # read alignment files... 
#> INFO  @ Sun, 13 Dec 2020 14:23:31: # read treatment tags... 
#> INFO  @ Sun, 13 Dec 2020 14:23:31: Detected format is: BED 
#> INFO  @ Sun, 13 Dec 2020 14:23:31: * Input file is gzipped. 
#> INFO  @ Sun, 13 Dec 2020 14:23:31: tag size is determined as 101 bps 
#> INFO  @ Sun, 13 Dec 2020 14:23:31: # tag size = 101 
#> INFO  @ Sun, 13 Dec 2020 14:23:31: # total tags in alignment file: 49622 
#> INFO  @ Sun, 13 Dec 2020 14:23:31: # Build Peak Model... 
#> INFO  @ Sun, 13 Dec 2020 14:23:31: #2 looking for paired plus/minus strand peaks... 
#> INFO  @ Sun, 13 Dec 2020 14:23:31: #2 Total number of paired peaks: 1031 
#> INFO  @ Sun, 13 Dec 2020 14:23:31: #2 Model building with cross-correlation: Done 
#> INFO  @ Sun, 13 Dec 2020 14:23:31: # finished! 
#> INFO  @ Sun, 13 Dec 2020 14:23:31: # predicted fragment length is 239 bps 
#> INFO  @ Sun, 13 Dec 2020 14:23:31: # alternative fragment length(s) may be 230,239 bps 
#> INFO  @ Sun, 13 Dec 2020 14:23:31: # Generate R script for model : /var/folders/pv/qm6yts0j7hdgggvq53hrzfr0vfyhvv/T//RtmpsaYg38/filefe1c28ba40bf 
#> 

#> Error in pdf("/var/folders/pv/qm6yts0j7hdgggvq53hrzfr0vfyhvv/T//RtmpsaYg38/filefe1c28ba40bf_model.pdf",     height = 6, width = 6): could not find function "pdf"