bdgbroadcall

Call broad peaks from bedGraph output. Note: All regions on the same chromosome in the bedGraph file should be continuous so only bedGraph files from MACS3 are accpetable.

bdgbroadcall(
  ifile,
  cutoffpeak = 2,
  cutofflink = 1,
  minlen = 200L,
  lvl1maxgap = 30L,
  lvl2maxgap = 800L,
  trackline = TRUE,
  outdir = ".",
  outputfile = character(),
  log = TRUE
)

Arguments

ifile
MACS score in bedGraph. REQUIRED.

cutoffpeak
Cutoff for peaks depending on which method you used for score track. If the file contains qvalue scores from MACS3, score 2 means qvalue 0.01. DEFAULT: 2

cutofflink
Cutoff for linking regions/low abundance regions depending on which method you used for score track. If the file contains qvalue scores from MACS3, score 1 means qvalue 0.1, and score 0.3 means qvalue 0.5. DEFAULT: 1", default = 1

minlen
minimum length of peak, better to set it as d value. DEFAULT: 200", default = 200

lvl1maxgap
maximum gap between significant peaks, better to set it as tag size. DEFAULT: 30

lvl2maxgap
maximum linking between significant peaks, better to set it as 4 times of d value. DEFAULT: 800

trackline
Tells MACS not to include trackline with bedGraph files. The trackline is required by UCSC.

outdir
The output directory.

outputfile
The output file.

log
Whether to capture logs.

Value

macsList object.

Examples

eh <- ExperimentHub::ExperimentHub()
CHIP <- eh[["EH4558"]]
CTRL <- eh[["EH4563"]]
p1 <- pileup(CHIP, outdir = tempdir(),
outputfile = "pileup_ChIP_bed.bdg", format = "BED")
p2 <- pileup(CTRL, outdir = tempdir(),
outputfile = "pileup_CTRL_bed.bdg", format = "BED")
c1 <- bdgcmp(p1$outputs, p2$outputs, outdir = tempdir(),
oprefix = "bdgcmp", pseudocount = 1, method = "FE")
bdgbroadcall(c1$outputs, cutoffpeak = 2, cutofflink = 1.5,
outdir = tempdir(), outputfile = "bdgbroadcall")