bdgopt

Operations on score column of bedGraph file. Note: All regions on the same chromosome in the bedGraph file should be continuous so only bedGraph files from MACS3 are accpetable.

bdgopt(
  ifile,
  method = c("multiply", "add", "p2q", "max", "min"),
  extraparam = numeric(),
  outputfile = character(),
  outdir = ".",
  log = TRUE,
  verbose = 2L
)

Arguments

ifile

MACS score in bedGraph. Note: this must be a bedGraph file covering the ENTIRE genome. REQUIRED

method

Method to modify the score column of bedGraph file. Available choices are: multiply, add, max, min, or p2q. 1) multiply, the EXTRAPARAM is required and will be multiplied to the score column. If you intend to divide the score column by X, use value of 1/X as EXTRAPARAM. 2) add, the EXTRAPARAM is required and will be added to the score column. If you intend to subtract the score column by X, use value of -X as EXTRAPARAM. 3) max, the EXTRAPARAM is required and will take the maximum value between score and the EXTRAPARAM. 4) min, the EXTRAPARAM is required and will take the minimum value between score and the EXTRAPARAM. 5) p2q, this will convert p-value scores to q-value scores using Benjamini-Hochberg process. The EXTRAPARAM is not required. This method assumes the scores are -log10 p-value from MACS3. Any other types of score will cause unexpected errors.", default="p2q"

extraparam

The extra parameter for METHOD. Check the detail of -m option.

outputfile

Output filename. Mutually exclusive with --o-prefix. The number and the order of arguments for --ofile must be the same as for -m.

outdir

The output directory.

log

Whether to capture logs.

verbose

Set verbose level of runtime message. 0: only show critical message, 1: show additional warning message, 2: show process information, 3: show debug messages. DEFAULT:2

Value

macsList object.

Examples

eh <- ExperimentHub::ExperimentHub()
CHIP <- eh[["EH4558"]]
#> see ?MACSdata and browseVignettes('MACSdata') for documentation
#> loading from cache
CTRL <- eh[["EH4563"]]
#> see ?MACSdata and browseVignettes('MACSdata') for documentation
#> loading from cache
c1 <- callpeak(CHIP, CTRL, gsize = 5.2e7, cutoff_analysis = TRUE,
               outdir = tempdir(), name = "callpeak_narrow0",
               store_bdg = TRUE)
#> INFO  @ 17 Nov 2023 13:42:18: [758 MB] 
#> # Command line: 
#> # ARGUMENTS LIST:
#> # name = callpeak_narrow0
#> # format = AUTO
#> # ChIP-seq file = ['/home/qhu/.cache/R/ExperimentHub/1071e4528bfa6_4601']
#> # control file = ['/home/qhu/.cache/R/ExperimentHub/1071e71fe08e5_4606']
#> # effective genome size = 5.20e+07
#> # band width = 300
#> # model fold = [5.0, 50.0]
#> # qvalue cutoff = 5.00e-02
#> # The maximum gap between significant sites is assigned as the read length/tag size.
#> # The minimum length of peaks is assigned as the predicted fragment length "d".
#> # Larger dataset will be scaled towards smaller dataset.
#> # Range for calculating regional lambda is: 1000 bps and 10000 bps
#> # Broad region calling is off
#> # Additional cutoff on fold-enrichment is: 0.10
#> # Paired-End mode is off
#>  
#> INFO  @ 17 Nov 2023 13:42:18: [758 MB] #1 read tag files... 
#> INFO  @ 17 Nov 2023 13:42:18: [758 MB] #1 read treatment tags... 
#> INFO  @ 17 Nov 2023 13:42:18: [758 MB] Detected format is: BED 
#> INFO  @ 17 Nov 2023 13:42:18: [758 MB] * Input file is gzipped. 
#> INFO  @ 17 Nov 2023 13:42:18: [758 MB] #1.2 read input tags... 
#> INFO  @ 17 Nov 2023 13:42:18: [758 MB] Detected format is: BED 
#> INFO  @ 17 Nov 2023 13:42:18: [758 MB] * Input file is gzipped. 
#> INFO  @ 17 Nov 2023 13:42:18: [758 MB] #1 tag size is determined as 101 bps 
#> INFO  @ 17 Nov 2023 13:42:18: [758 MB] #1 tag size = 101.0 
#> INFO  @ 17 Nov 2023 13:42:18: [758 MB] #1  total tags in treatment: 49622 
#> INFO  @ 17 Nov 2023 13:42:18: [758 MB] #1 user defined the maximum tags... 
#> INFO  @ 17 Nov 2023 13:42:18: [758 MB] #1 filter out redundant tags at the same location and the same strand by allowing at most 1 tag(s) 
#> INFO  @ 17 Nov 2023 13:42:18: [758 MB] #1  tags after filtering in treatment: 48047 
#> INFO  @ 17 Nov 2023 13:42:18: [758 MB] #1  Redundant rate of treatment: 0.03 
#> INFO  @ 17 Nov 2023 13:42:18: [758 MB] #1  total tags in control: 50837 
#> INFO  @ 17 Nov 2023 13:42:18: [758 MB] #1 user defined the maximum tags... 
#> INFO  @ 17 Nov 2023 13:42:18: [758 MB] #1 filter out redundant tags at the same location and the same strand by allowing at most 1 tag(s) 
#> INFO  @ 17 Nov 2023 13:42:18: [758 MB] #1  tags after filtering in control: 50783 
#> INFO  @ 17 Nov 2023 13:42:18: [758 MB] #1  Redundant rate of control: 0.00 
#> INFO  @ 17 Nov 2023 13:42:18: [758 MB] #1 finished! 
#> INFO  @ 17 Nov 2023 13:42:18: [758 MB] #2 Build Peak Model... 
#> INFO  @ 17 Nov 2023 13:42:18: [758 MB] #2 looking for paired plus/minus strand peaks... 
#> INFO  @ 17 Nov 2023 13:42:18: [758 MB] #2 Total number of paired peaks: 469 
#> INFO  @ 17 Nov 2023 13:42:18: [758 MB] #2 Model building with cross-correlation: Done 
#> INFO  @ 17 Nov 2023 13:42:18: [758 MB] #2 finished! 
#> INFO  @ 17 Nov 2023 13:42:18: [758 MB] #2 predicted fragment length is 228 bps 
#> INFO  @ 17 Nov 2023 13:42:18: [758 MB] #2 alternative fragment length(s) may be 228 bps 
#> INFO  @ 17 Nov 2023 13:42:18: [758 MB] #2.2 Generate R script for model : /tmp/RtmpEPj3Vo/callpeak_narrow0_model.r 
#> INFO  @ 17 Nov 2023 13:42:18: [758 MB] #3 Call peaks... 
#> INFO  @ 17 Nov 2023 13:42:18: [758 MB] #3 Pre-compute pvalue-qvalue table... 
#> INFO  @ 17 Nov 2023 13:42:18: [758 MB] #3 Cutoff vs peaks called will be analyzed! 
#> INFO  @ 17 Nov 2023 13:42:19: [758 MB] #3 Analysis of cutoff vs num of peaks or total length has been saved in b'/tmp/RtmpEPj3Vo/callpeak_narrow0_cutoff_analysis.txt' 
#> INFO  @ 17 Nov 2023 13:42:19: [758 MB] #3 In the peak calling step, the following will be performed simultaneously: 
#> INFO  @ 17 Nov 2023 13:42:19: [758 MB] #3   Write bedGraph files for treatment pileup (after scaling if necessary)... callpeak_narrow0_treat_pileup.bdg 
#> INFO  @ 17 Nov 2023 13:42:19: [758 MB] #3   Write bedGraph files for control lambda (after scaling if necessary)... callpeak_narrow0_control_lambda.bdg 
#> INFO  @ 17 Nov 2023 13:42:19: [758 MB] #3   Pileup will be based on sequencing depth in treatment. 
#> INFO  @ 17 Nov 2023 13:42:19: [758 MB] #3 Call peaks for each chromosome... 
#> INFO  @ 17 Nov 2023 13:42:19: [758 MB] #4 Write output xls file... /tmp/RtmpEPj3Vo/callpeak_narrow0_peaks.xls 
#> INFO  @ 17 Nov 2023 13:42:19: [758 MB] #4 Write peak in narrowPeak format file... /tmp/RtmpEPj3Vo/callpeak_narrow0_peaks.narrowPeak 
#> INFO  @ 17 Nov 2023 13:42:19: [758 MB] #4 Write summits bed file... /tmp/RtmpEPj3Vo/callpeak_narrow0_summits.bed 
#> INFO  @ 17 Nov 2023 13:42:19: [758 MB] Done! 
#> 
cfile <- grep("treat_pileup.bdg", c1$outputs, value = TRUE)
bdgopt(cfile, method = "min", extraparam = 10,
       outdir = tempdir(), outputfile = "bdgopt_min.bdg")
#> INFO  @ 17 Nov 2023 13:42:19: [758 MB] Read and build bedGraph... 
#> INFO  @ 17 Nov 2023 13:42:19: [758 MB] Modify bedGraph... 
#> INFO  @ 17 Nov 2023 13:42:19: [758 MB] Write bedGraph of modified scores... 
#> INFO  @ 17 Nov 2023 13:42:19: [758 MB] Finished 'min'! Please check '/tmp/RtmpEPj3Vo/bdgopt_min.bdg'! 
#> 
#> macsList class
#> $outputs:
#>  /tmp/RtmpEPj3Vo/bdgopt_min.bdg 
#> $arguments: ifile, method, extraparam, outputfile, outdir 
#> $log:
#>  INFO  @ 17 Nov 2023 13:42:19: [758 MB] Read and build bedGraph... 
#>  INFO  @ 17 Nov 2023 13:42:19: [758 MB] Modify bedGraph... 
#>  INFO  @ 17 Nov 2023 13:42:19: [758 MB] Write bedGraph of modified scores... 
#>  INFO  @ 17 Nov 2023 13:42:19: [758 MB] Finished 'min'! Please check '/tmp/RtmpEPj3Vo/bdgopt_min.bdg'! 
#>