predictd

predictd

predictd(
  ifile,
  gsize = "hs",
  format = "AUTO",
  plot = file.path(tempdir(), "predictd_mode.pdf"),
  tsize = NULL,
  bw = 300,
  d_min = 20,
  mfold = c(5, 50),
  buffer_size = 1e+05,
  verbose = 2L,
  log = TRUE
)

Arguments

ifile
Input file(s).

gsize
Effective genome size. It can be 1.0e+9 or 1000000000, or shortcuts:'hs' for human (2.7e9), 'mm' for mouse (1.87e9), 'ce' for C. elegans (9e7) and 'dm' for fruitfly (1.2e8), Default:hs.

format
Input file format.

plot
PDF path of peak model and correlation plots.

tsize
Tag size. This will override the auto detected tag size.

bw
Band width for picking regions to compute fragment size. This value is only used while building the shifting model. DEFAULT: 300

d_min
Minimum fragment size in basepair. Any predicted fragment size less than this will be excluded. DEFAULT: 20

mfold
Select the regions within MFOLD range of high-confidence enrichment ratio against background to build model. Fold-enrichment in regions must be lower than upper limit, and higher than the lower limit. Use as "-m 10 30". DEFAULT:5 50

buffer_size
Buffer size for incrementally increasing internal array size to store reads alignment information. DEFAULT: 100000.

verbose
Set verbose level of runtime message. 0: only show critical message, 1: show additional warning message, 2: show process information, 3: show debug messages. DEFAULT:2

log
Whether to capture log.

Examples

if (FALSE) {
predictd("extdata/CTCF_SE_ChIP_chr22_50k.bed.gz", d_min = 10)
}