randsample

Randomly sample number/percentage of total reads.

randsample(
  ifile,
  outdir = ".",
  outputfile = character(),
  percentage = numeric(),
  number = numeric(),
  seed = -1L,
  tsize = NULL,
  format = c("AUTO", "BAM", "SAM", "BED", "ELAND", "ELANDMULTI", "ELANDEXPORT", "BOWTIE",
    "BAMPE", "BEDPE"),
  buffer_size = 100000L,
  verbose = 2L,
  log = TRUE
)

Arguments

ifile

Alignment file. If multiple files are given as '-t A B C', then they will all be read and combined. Note that pair-end data is not supposed to work with this command. REQUIRED.

outdir

The output directory.

outputfile

Output bedGraph file name. If not specified, will write to standard output. REQUIRED.

percentage

Percentage of tags you want to keep. Input 80.0 for 80%%. This option can't be used at the same time with -n/--num. REQUIRED

number

Number of tags you want to keep. Input 8000000 or 8e+6 for 8 million. This option can't be used at the same time with -p/--percent. Note that the number of tags in output is approximate as the number specified here. REQUIRED

seed

Set the random seed while down sampling data. Must be a non-negative integer in order to be effective. DEFAULT: not set

tsize

Tag size. This will override the auto detected tag size. DEFAULT: Not set

format

Format of tag file, \"AUTO\", \"BED\" or \"ELAND\" or \"ELANDMULTI\" or \"ELANDEXPORT\" or \"SAM\" or \"BAM\" or \"BOWTIE\" or \"BAMPE\" or \"BEDPE\". The default AUTO option will %(prog)s decide which format the file is. Please check the definition in README file if you choose ELAND/ELANDMULTI/ELANDEXPORT/SAM/BAM/BOWTIE or BAMPE/BEDPE. DEFAULT: \"AUTO\""

buffer_size

Buffer size for incrementally increasing internal array size to store reads alignment information. In most cases, you don't have to change this parameter. However, if there are large number of chromosomes/contigs/scaffolds in your alignment, it's recommended to specify a smaller buffer size in order to decrease memory usage (but it will take longer time to read alignment files). Minimum memory requested for reading an alignment file is about # of CHROMOSOME * BUFFER_SIZE * 8 Bytes. DEFAULT: 100000

verbose

Set verbose level. 0: only show critical message, 1: show additional warning message, 2: show process information, 3: show debug messages. If you want to know where are the duplicate reads, use 3. DEFAULT:2

log

Whether to capture logs.

Value

macsList object.

Examples

eh <- ExperimentHub::ExperimentHub()
CHIP <- eh[["EH4558"]]
#> see ?MACSdata and browseVignettes('MACSdata') for documentation
#> loading from cache
randsample(CHIP, number = 1000, outdir = tempdir(), outputfile = "randsample.bed")
#> INFO  @ 17 Nov 2023 13:42:47: [761 MB] read tag files... 
#> INFO  @ 17 Nov 2023 13:42:47: [761 MB] # read treatment tags... 
#> INFO  @ 17 Nov 2023 13:42:47: [761 MB] Detected format is: BED 
#> INFO  @ 17 Nov 2023 13:42:47: [761 MB] * Input file is gzipped. 
#> INFO  @ 17 Nov 2023 13:42:47: [761 MB] tag size is determined as 101 bps 
#> INFO  @ 17 Nov 2023 13:42:47: [761 MB] tag size = 101 
#> INFO  @ 17 Nov 2023 13:42:47: [761 MB]  total tags in alignment file: 49622 
#> INFO  @ 17 Nov 2023 13:42:47: [761 MB]  Number of tags you want to keep: 1.00e+03 
#> INFO  @ 17 Nov 2023 13:42:47: [761 MB]  Percentage of tags you want to keep: 2.02% 
#> INFO  @ 17 Nov 2023 13:42:47: [761 MB]  tags after random sampling in alignment file: 999 
#> INFO  @ 17 Nov 2023 13:42:47: [761 MB] Write to BED file 
#> INFO  @ 17 Nov 2023 13:42:47: [761 MB] finished! Check randsample.bed. 
#> 
#> macsList class
#> $outputs:
#>  /tmp/RtmpEPj3Vo/randsample.bed 
#> $arguments: ifile, outdir, outputfile, number 
#> $log:
#>  INFO  @ 17 Nov 2023 13:42:47: [761 MB] read tag files... 
#>  INFO  @ 17 Nov 2023 13:42:47: [761 MB] # read treatment tags... 
#>  INFO  @ 17 Nov 2023 13:42:47: [761 MB] Detected format is: BED 
#>  INFO  @ 17 Nov 2023 13:42:47: [761 MB] * Input file is gzipped. 
#>  INFO  @ 17 Nov 2023 13:42:47: [761 MB] tag size is determined as 101 bps 
#> ...