refinepeak

(Experimental) Take raw reads alignment, refine peak summits and give scores measuring balance of waston/crick tags. Inspired by SPP.

refinepeak(
  bedfile,
  ifile,
  format = c("AUTO", "BAM", "SAM", "BED", "ELAND", "ELANDMULTI", "ELANDEXPORT", "BOWTIE"),
  cutoff = 5,
  windowsize = 200L,
  buffer_size = 100000L,
  verbose = 2L,
  outdir = "./",
  outputfile = character(),
  log = TRUE
)

Arguments

bedfile

Candidate peak file in BED format. REQUIRED.

ifile

ChIP-seq alignment file. If multiple files are given as '-t A B C', then they will all be read and combined. Note that pair-end data is not supposed to work with this command. REQUIRED.

format

Format of tag file, \"AUTO\", \"BED\" or \"ELAND\" or \"ELANDMULTI\" or \"ELANDEXPORT\" or \"SAM\" or \"BAM\" or \"BOWTIE\". The default AUTO option will let '%(prog)s' decide which format the file is. Please check the definition in README file if you choose ELAND/ELANDMULTI/ELANDEXPORT/SAM/BAM/BOWTIE. DEFAULT: \"AUTO\""

cutoff

Cutoff DEFAULT: 5

windowsize

Scan window size on both side of the summit (default: 100bp)

buffer_size

Buffer size for incrementally increasing internal array size to store reads alignment information. In most cases, you don't have to change this parameter. However, if there are large number of chromosomes/contigs/scaffolds in your alignment, it's recommended to specify a smaller buffer size in order to decrease memory usage (but it will take longer time to read alignment files). Minimum memory requested for reading an alignment file is about # of CHROMOSOME * BUFFER_SIZE * 8 Bytes. DEFAULT: 100000

verbose

Set verbose level. 0: only show critical message, 1: show additional warning message, 2: show process information, 3: show debug messages. If you want to know where are the duplicate reads, use 3. DEFAULT:2

outdir

The output directory.

outputfile

Output bedGraph file name. If not specified, will write to standard output. REQUIRED.

log

Whether to capture logs.

Value

macsList object.

Examples

eh <- ExperimentHub::ExperimentHub()
CHIP <- eh[["EH4558"]]
#> see ?MACSdata and browseVignettes('MACSdata') for documentation
#> loading from cache
CTRL <- eh[["EH4563"]]
#> see ?MACSdata and browseVignettes('MACSdata') for documentation
#> loading from cache
res <- callpeak(CHIP, CTRL, gsize = 5.2e7, cutoff_analysis = TRUE,
                outdir = tempdir(), name = "callpeak_narrow0")
#> INFO  @ 17 Nov 2023 12:41:45: [763 MB] 
#> # Command line: 
#> # ARGUMENTS LIST:
#> # name = callpeak_narrow0
#> # format = AUTO
#> # ChIP-seq file = ['/home/qhu/.cache/R/ExperimentHub/1071e4528bfa6_4601']
#> # control file = ['/home/qhu/.cache/R/ExperimentHub/1071e71fe08e5_4606']
#> # effective genome size = 5.20e+07
#> # band width = 300
#> # model fold = [5.0, 50.0]
#> # qvalue cutoff = 5.00e-02
#> # The maximum gap between significant sites is assigned as the read length/tag size.
#> # The minimum length of peaks is assigned as the predicted fragment length "d".
#> # Larger dataset will be scaled towards smaller dataset.
#> # Range for calculating regional lambda is: 1000 bps and 10000 bps
#> # Broad region calling is off
#> # Additional cutoff on fold-enrichment is: 0.10
#> # Paired-End mode is off
#>  
#> INFO  @ 17 Nov 2023 12:41:45: [763 MB] #1 read tag files... 
#> INFO  @ 17 Nov 2023 12:41:45: [763 MB] #1 read treatment tags... 
#> INFO  @ 17 Nov 2023 12:41:45: [763 MB] Detected format is: BED 
#> INFO  @ 17 Nov 2023 12:41:45: [763 MB] * Input file is gzipped. 
#> INFO  @ 17 Nov 2023 12:41:45: [763 MB] #1.2 read input tags... 
#> INFO  @ 17 Nov 2023 12:41:45: [763 MB] Detected format is: BED 
#> INFO  @ 17 Nov 2023 12:41:45: [763 MB] * Input file is gzipped. 
#> INFO  @ 17 Nov 2023 12:41:45: [763 MB] #1 tag size is determined as 101 bps 
#> INFO  @ 17 Nov 2023 12:41:45: [763 MB] #1 tag size = 101.0 
#> INFO  @ 17 Nov 2023 12:41:45: [763 MB] #1  total tags in treatment: 49622 
#> INFO  @ 17 Nov 2023 12:41:45: [763 MB] #1 user defined the maximum tags... 
#> INFO  @ 17 Nov 2023 12:41:45: [763 MB] #1 filter out redundant tags at the same location and the same strand by allowing at most 1 tag(s) 
#> INFO  @ 17 Nov 2023 12:41:45: [763 MB] #1  tags after filtering in treatment: 48047 
#> INFO  @ 17 Nov 2023 12:41:45: [763 MB] #1  Redundant rate of treatment: 0.03 
#> INFO  @ 17 Nov 2023 12:41:45: [763 MB] #1  total tags in control: 50837 
#> INFO  @ 17 Nov 2023 12:41:45: [763 MB] #1 user defined the maximum tags... 
#> INFO  @ 17 Nov 2023 12:41:45: [763 MB] #1 filter out redundant tags at the same location and the same strand by allowing at most 1 tag(s) 
#> INFO  @ 17 Nov 2023 12:41:45: [763 MB] #1  tags after filtering in control: 50783 
#> INFO  @ 17 Nov 2023 12:41:45: [763 MB] #1  Redundant rate of control: 0.00 
#> INFO  @ 17 Nov 2023 12:41:45: [763 MB] #1 finished! 
#> INFO  @ 17 Nov 2023 12:41:45: [763 MB] #2 Build Peak Model... 
#> INFO  @ 17 Nov 2023 12:41:45: [763 MB] #2 looking for paired plus/minus strand peaks... 
#> INFO  @ 17 Nov 2023 12:41:45: [763 MB] #2 Total number of paired peaks: 469 
#> INFO  @ 17 Nov 2023 12:41:45: [763 MB] #2 Model building with cross-correlation: Done 
#> INFO  @ 17 Nov 2023 12:41:45: [763 MB] #2 finished! 
#> INFO  @ 17 Nov 2023 12:41:45: [763 MB] #2 predicted fragment length is 228 bps 
#> INFO  @ 17 Nov 2023 12:41:45: [763 MB] #2 alternative fragment length(s) may be 228 bps 
#> INFO  @ 17 Nov 2023 12:41:45: [763 MB] #2.2 Generate R script for model : /tmp/RtmpeMHyqk/callpeak_narrow0_model.r 
#> INFO  @ 17 Nov 2023 12:41:45: [763 MB] #3 Call peaks... 
#> INFO  @ 17 Nov 2023 12:41:45: [763 MB] #3 Pre-compute pvalue-qvalue table... 
#> INFO  @ 17 Nov 2023 12:41:45: [763 MB] #3 Cutoff vs peaks called will be analyzed! 
#> INFO  @ 17 Nov 2023 12:41:46: [769 MB] #3 Analysis of cutoff vs num of peaks or total length has been saved in b'/tmp/RtmpeMHyqk/callpeak_narrow0_cutoff_analysis.txt' 
#> INFO  @ 17 Nov 2023 12:41:46: [769 MB] #3 Call peaks for each chromosome... 
#> INFO  @ 17 Nov 2023 12:41:46: [770 MB] #4 Write output xls file... /tmp/RtmpeMHyqk/callpeak_narrow0_peaks.xls 
#> INFO  @ 17 Nov 2023 12:41:46: [770 MB] #4 Write peak in narrowPeak format file... /tmp/RtmpeMHyqk/callpeak_narrow0_peaks.narrowPeak 
#> INFO  @ 17 Nov 2023 12:41:46: [770 MB] #4 Write summits bed file... /tmp/RtmpeMHyqk/callpeak_narrow0_summits.bed 
#> INFO  @ 17 Nov 2023 12:41:46: [770 MB] Done! 
#> 
refinepeak(grep("narrowPeak", res$outputs, value = TRUE), CHIP,
           outdir = tempdir(), outputfile = "refine")
#> INFO  @ 17 Nov 2023 12:41:46: [770 MB] read tag files... 
#> INFO  @ 17 Nov 2023 12:41:46: [770 MB] # read treatment tags... 
#> INFO  @ 17 Nov 2023 12:41:46: [770 MB] Detected format is: BED 
#> INFO  @ 17 Nov 2023 12:41:46: [770 MB] * Input file is gzipped. 
#> INFO  @ 17 Nov 2023 12:41:47: [770 MB] Done! 
#> 
#> macsList class
#> $outputs:
#>  /tmp/RtmpeMHyqk/refine 
#> $arguments: bedfile, ifile, outdir, outputfile 
#> $log:
#>  INFO  @ 17 Nov 2023 12:41:46: [770 MB] read tag files... 
#>  INFO  @ 17 Nov 2023 12:41:46: [770 MB] # read treatment tags... 
#>  INFO  @ 17 Nov 2023 12:41:46: [770 MB] Detected format is: BED 
#>  INFO  @ 17 Nov 2023 12:41:46: [770 MB] * Input file is gzipped. 
#>  INFO  @ 17 Nov 2023 12:41:47: [770 MB] Done! 
#> ...